sample/requirements.txt
asgiref==3.4.1
Django==3.2.8
djangorestframework==3.12.4
html2text==2020.1.16
Jinja2==3.0.2
ldap3==2.9.1
MarkupSafe==2.0.1
pyasn1==0.4.8
pytz==2021.3
PyYAML==5.4.1
sqlparse==0.4.2
